backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.0.0
aiofiles>=23.0.0
